+++ pkg/spi-shared/tokenstorage/tokenstorage.go
@@ -16,10 +16,6 @@ package tokenstorage
 
 import (
 	"context"
-	"fmt"
-
-	vault "github.com/hashicorp/vault/api"
-	auth "github.com/hashicorp/vault/api/auth/kubernetes"
 
 	api "github.com/redhat-appstudio/service-provider-integration-operator/api/v1beta1"
 )
@@ -31,32 +27,4 @@ type TokenStorage interface {
 	Get(ctx context.Context, owner *api.SPIAccessToken) (*api.Token, error)
 	GetDataLocation(ctx context.Context, owner *api.SPIAccessToken) (string, error)
 	Delete(ctx context.Context, owner *api.SPIAccessToken) error
-}
-
-// NewVaultStorage creates a new `TokenStorage` instance using the provided Kubernetes client.
-func NewVaultStorage(role string, vaultHost string, serviceAccountToken string) (TokenStorage, error) {
-	config := vault.DefaultConfig()
-	config.Address = vaultHost
-	client, err := vault.NewClient(config)
-	if err != nil {
-		return nil, err
-	}
-	var k8sAuth *auth.KubernetesAuth
-	if serviceAccountToken == "" {
-		k8sAuth, err = auth.NewKubernetesAuth(role)
-	} else {
-		k8sAuth, err = auth.NewKubernetesAuth(role, auth.WithServiceAccountTokenPath(serviceAccountToken))
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	authInfo, err := client.Auth().Login(context.TODO(), k8sAuth)
-	if err != nil {
-		return nil, err
-	}
-	if authInfo == nil {
-		return nil, fmt.Errorf("no auth info was returned after login to vault")
-	}
-	return &vaultTokenStorage{client}, nil
 }
